Split HTML renderer & document it
