+++ crates/rust-glancer/src/analyze/report/render/html/assets/report.js
